requirements.txt
flask==2.2.2
pillow==9.3.0